+++ src/main/kotlin/kotools/types/number/NegativeInt.kt
@@ -56,7 +56,7 @@ public fun Int.toNegativeIntOrNull(): NegativeInt? = NegativeInt orNull this
 @SinceKotoolsTypes("1.1")
 public value class NegativeInt
 @Throws(IllegalArgumentException::class)
-public constructor(public val value: Int) : Comparable<NegativeInt> {
+public constructor(public val value: Int) : Comparable<Int> {
     init {
         require(value <= 0) { "Given value shouldn't be strictly positive." }
     }
@@ -209,8 +209,7 @@ public constructor(public val value: Int) : Comparable<NegativeInt> {
      * this [value] is greater than the [other] value.
      */
     @SinceKotoolsTypes("1.3")
-    public infix operator fun compareTo(other: Int): Int =
-        value.compareTo(other)
+    override infix fun compareTo(other: Int): Int = value.compareTo(other)
 
     /**
      * Compares this [value] with the [other] value for order.
@@ -245,7 +244,7 @@ public constructor(public val value: Int) : Comparable<NegativeInt> {
      * if this [value] is less than the [other] value, or a positive number if
      * this [value] is greater than the [other] value.
      */
-    override infix fun compareTo(other: NegativeInt): Int =
+    public infix operator fun compareTo(other: NegativeInt): Int =
         compareTo(other.value)
 
     /**