+++ src/main/kotlin/kotools/types/number/NonZeroInt.kt
@@ -64,7 +64,7 @@ public fun Int.toNonZeroIntOrNull(): NonZeroInt? = NonZeroInt orNull this
 @SinceKotoolsTypes("1.1")
 public value class NonZeroInt
 @Throws(IllegalArgumentException::class)
-public constructor(public val value: Int) : Comparable<NonZeroInt> {
+public constructor(public val value: Int) : Comparable<Int> {
     init {
         require(value != 0) { "Given value shouldn't equal 0." }
     }
@@ -221,8 +221,7 @@ public constructor(public val value: Int) : Comparable<NonZeroInt> {
      * this [value] is greater than the [other] value.
      */
     @SinceKotoolsTypes("1.3")
-    public infix operator fun compareTo(other: Int): Int =
-        value.compareTo(other)
+    override infix fun compareTo(other: Int): Int = value.compareTo(other)
 
     /**
      * Compares this [value] with the [other] value for order.
@@ -230,7 +229,7 @@ public constructor(public val value: Int) : Comparable<NonZeroInt> {
      * if this [value] is less than the [other] value, or a positive number if
      * this [value] is greater than the [other] value.
      */
-    override infix operator fun compareTo(other: NonZeroInt): Int =
+    public infix operator fun compareTo(other: NonZeroInt): Int =
         compareTo(other.value)
 
     /**