+++ src/main/kotlin/kotools/types/number/StrictlyNegativeInt.kt
@@ -57,7 +57,7 @@ public fun Int.toStrictlyNegativeIntOrNull(): StrictlyNegativeInt? =
 @SinceKotoolsTypes("1.1")
 public value class StrictlyNegativeInt
 @Throws(IllegalArgumentException::class)
-public constructor(public val value: Int) : Comparable<StrictlyNegativeInt> {
+public constructor(public val value: Int) : Comparable<Int> {
     init {
         require(value < 0) { "Given value shouldn't be positive." }
     }
@@ -214,8 +214,7 @@ public constructor(public val value: Int) : Comparable<StrictlyNegativeInt> {
      * this [value] is greater than the [other] value.
      */
     @SinceKotoolsTypes("1.3")
-    public infix operator fun compareTo(other: Int): Int =
-        value.compareTo(other)
+    override infix fun compareTo(other: Int): Int = value.compareTo(other)
 
     /**
      * Compares this [value] with the [other] value for order.
@@ -259,7 +258,7 @@ public constructor(public val value: Int) : Comparable<StrictlyNegativeInt> {
      * if this [value] is less than the [other] value, or a positive number if
      * this [value] is greater than the [other] value.
      */
-    override infix fun compareTo(other: StrictlyNegativeInt): Int =
+    public infix operator fun compareTo(other: StrictlyNegativeInt): Int =
         compareTo(other.value)
 
     // ---------- Conversions ----------